serviceの取得処理を切り出し
