+++ lib/google_spreadsheet_fetcher.rb
@@ -4,6 +4,7 @@ require "google_spreadsheet_fetcher/version"
 require "google_spreadsheet_fetcher/config"
 require "google_spreadsheet_fetcher/error"
 require "google_spreadsheet_fetcher/fetcher"
+require "google_spreadsheet_fetcher/sheets_service_builder"
 
 module GoogleSpreadsheetFetcher
   def self.config
